+++ backend/data/notification.data.js
@@ -37,8 +37,9 @@ class Notification {
      * @param {function(Error, Notification[]):void} callback - La fonction de rappel à exécuter après la récupération des notifications.
      */
     static getAllNotifications(callback) {
-        const query =
-            "SELECT * FROM notifications JOIN employees ON notifications.id_employee = employees.id_employee ORDER BY creation_date DESC, creation_hour DESC";
+        const query = `SELECT * FROM notifications JOIN employees ON notifications.id_employee = employees.id_employee 
+            JOIN events ON notifications.id_employee = events.id_employee
+            ORDER BY creation_date DESC, creation_hour DESC`;
         pool.query(query, (error, result) => {
             if (error) {
                 return callback(error, null);
@@ -55,6 +56,7 @@ class Notification {
                 );
                 notification.firstName = row.firstname;
                 notification.lastName = row.lastname;
+                notification.idEvent = row.id_event;
                 return notification;
             });
             callback(null, notifications);