+++ frontend/src/components/InterventionForm/InterventionForm.js
@@ -316,7 +316,7 @@ class InterventionForm extends Component {
                                                                 .handleViewReport
                                                         }
                                                     >
-                                                        <i class="fa-solid fa-file-circle-check"></i>{" "}
+                                                        <i className="fa-solid fa-file-circle-check"></i>{" "}
                                                         Voir le rapport validé
                                                     </button>
                                                 );
@@ -530,7 +530,7 @@ class InterventionForm extends Component {
                                                                 .handleViewReport
                                                         }
                                                     >
-                                                        <i class="fa-solid fa-file-circle-check"></i>
+                                                        <i className="fa-solid fa-file-circle-check"></i>
                                                         {""}
                                                         Voir le rapport validé
                                                     </button>